+++ fragility.py
@@ -31,6 +31,12 @@ SUPPORTED_FRAGILITY_FUNCTION_FACTORIES = {
 class DamageState:
     '''
     Class to represent the damage states.
+
+    There are some attributes that are not specific for
+    damage states as the intensity_field, the intensity_unit,
+    the kind of fragility function (lognormcdf for example) or
+    the min and max intensities. However in order to work with
+    those values more easily, they are included here.
     '''
     def __init__(self,
                  taxonomy,
@@ -38,7 +44,9 @@ class DamageState:
                  to_state,
                  intensity_field,
                  intensity_unit,
-                 fragility_function):
+                 fragility_function,
+                 max_intensity=np.inf,
+                 min_intensity=-1*np.inf):
         self.taxonomy = taxonomy
         self.from_state = from_state
         self.to_state = to_state
@@ -46,6 +54,9 @@ class DamageState:
         self.intensity_unit = intensity_unit
 
         self.fragility_function = fragility_function
+
+        self.max_intensity = max_intensity
+        self.min_intensity = min_intensity
 
     def get_probability_for_intensity(self, intensity, units):
         '''
@@ -71,6 +82,11 @@ class DamageState:
 
         if unit != self.intensity_unit:
             raise Exception('Not supported unit')
+
+        if value < self.min_intensity:
+            return 0
+        if value > self.max_intensity:
+            return 1
 
         return self.fragility_function(value)
 
@@ -109,6 +125,15 @@ class Fragility:
             taxonomy = dataset['taxonomy']
             intensity_field = dataset['imt']
             intensity_unit = dataset['imu']
+
+            min_intensity = -1 * np.inf
+            max_intensity = np.inf
+
+            if 'im_min' in dataset.keys():
+                min_intensity = dataset['im_min']
+            if 'im_max' in dataset.keys():
+                max_intensity = dataset['im_max']
+
             for damage_state_mean_key in [
                     k for k in dataset.keys()
                     if k.startswith('D')
@@ -148,7 +173,9 @@ class Fragility:
                     to_state=to_state,
                     intensity_field=intensity_field,
                     intensity_unit=intensity_unit,
-                    fragility_function=fragility_function(mean, stddev)
+                    fragility_function=fragility_function(mean, stddev),
+                    max_intensity=max_intensity,
+                    min_intensity=min_intensity,
                 )
 
                 damage_states_by_taxonomy[taxonomy].append(damage_state)
@@ -215,6 +242,8 @@ class Fragility:
                             intensity_field=ds_lower.intensity_field,
                             intensity_unit=ds_lower.intensity_unit,
                             fragility_function=ds_lower.fragility_function,
+                            max_intensity=ds_lower.max_intensity,
+                            min_intensity=ds_lower.min_intensity,
                         )
                         damage_states.append(ds_new)